+++ ee/packages/presence/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @rocket.chat/presence
 
+## 0.1.0-rc.7
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.7
+- @rocket.chat/core-services@0.3.0-rc.7
+- @rocket.chat/models@0.0.21-rc.7
+
+## 0.1.0-rc.6
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.6
+- @rocket.chat/core-services@0.3.0-rc.6
+- @rocket.chat/models@0.0.21-rc.6
+
+## 0.1.0-rc.5
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.5
+- @rocket.chat/core-services@0.3.0-rc.5
+- @rocket.chat/models@0.0.21-rc.5
+
+## 0.1.0-rc.4
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.4
+- @rocket.chat/core-services@0.3.0-rc.4
+- @rocket.chat/models@0.0.21-rc.4
+
+## 0.1.0-rc.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.0-rc.3
+- @rocket.chat/core-services@0.3.0-rc.3
+- @rocket.chat/models@0.0.21-rc.3
+
 ## 0.1.0-rc.2
 
 ### Patch Changes
@@ -37,6 +77,30 @@
   - @rocket.chat/core-typings@6.5.0-rc.0
   - @rocket.chat/core-services@0.3.0-rc.0
   - @rocket.chat/models@0.0.21-rc.0
+
+## 0.0.23
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.8
+- @rocket.chat/core-services@0.2.8
+- @rocket.chat/models@0.0.23
+
+## 0.0.22
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.7
+- @rocket.chat/core-services@0.2.7
+- @rocket.chat/models@0.0.22
+
+## 0.0.21
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.4.6
+- @rocket.chat/core-services@0.2.6
+- @rocket.chat/models@0.0.21
 
 ## 0.0.20